完成前后端对接和测试

+++ Client/CodeBaseBuild/build_codebase.py
@@ -2,6 +2,8 @@ import os
 import json
 import re
 import glob
+import shutil
+import stat
 
 from git import Repo
 from copy import deepcopy
@@ -13,23 +15,28 @@ from client.CodeBaseBuild.llm_gen import generate_api
 
 import pandas as pd
 
-def repo_parse(repo_path, lib_path, output_path):
+def repo_parse(repo_path, lib_path, output_path, version):
     
-    c_files = glob.glob(f"{repo_path}/**/*.c", recursive=True)
-    h_files = glob.glob(f"{repo_path}/**/*.h", recursive=True)
+    if os.path.exists(output_path):
+        result = (f'代码库{os.path.basename(repo_path)}_{version}已存在，跳过解析步骤，仅进行数据解析。')
+    else:
+        c_files = glob.glob(f"{repo_path}/**/*.c", recursive=True)
+        h_files = glob.glob(f"{repo_path}/**/*.h", recursive=True)
 
-    c_parser = CParser(lib_path)
-    all_functions = []
+        c_parser = CParser(lib_path)
+        all_functions = []
 
-    for index, c_file in enumerate((c_files + h_files)):
-        print(f'processing{index+1}/{len((c_files + h_files))}:{c_file}')
-        c_parser.parse_file(c_file)
-        for func in c_parser.functions:
-            func["file_path"] = os.path.relpath(c_file, repo_path)
-        all_functions.extend(deepcopy(c_parser.functions))
-    df = pd.DataFrame(all_functions)
-    df.to_csv(output_path, index=False, encoding='utf-8-sig')
-    return f'成功建立代码库'
+        for index, c_file in enumerate((c_files + h_files)):
+            print(f'processing{index+1}/{len((c_files + h_files))}:{c_file}')
+            c_parser.parse_file(c_file)
+            for func in c_parser.functions:
+                func["file_path"] = os.path.relpath(c_file, repo_path)
+            all_functions.extend(deepcopy(c_parser.functions))
+        df = pd.DataFrame(all_functions)
+        df.to_csv(output_path, index=False, encoding='utf-8-sig')
+        result = f'代码库{os.path.basename(repo_path)}_{version}解析完成。'
+    rm_repo(repo_path)
+    return result
 
 def gen_sum(codebase_path):
 
@@ -38,7 +45,7 @@ def gen_sum(codebase_path):
     codebase = pd.read_csv(codebase_path)
 
     for index, row in codebase.iterrows():
-        
+        print(f'processing{index+1}/{len((codebase))}:{row["signature"]}')
         if (row['summary'] is not None) and (row['summary'] != "Not Generated") and (len(row['summary']) < 300):
             continue
         param = {'name': row['name'], 'file_path': row['file_path'], 'source_code': row['source_code'], 'signature': row['signature']}
@@ -80,18 +87,24 @@ def gen_sum(codebase_path):
             ite += 1
             if ite > 10:
                 break
-        print(f'【{index}/{len(codebase)}】【提示词】:{messages} \n 【回复】:{summary}')
+        # print(f'【{index}/{len(codebase)}】【提示词】:{messages} \n 【回复】:{summary}')
         if not summary: summary = "摘要失败."
         row['summary'] = summary.strip()
     codebase.to_csv(codebase_path, index=True, encoding='utf-8')
+    return f'成功生成代码库摘要'
 
-def sum_tokenize(codebase_path):
+def sum_embedding(codebase_path):
     codebase = pd.read_csv(codebase_path)
     codebase['sum_embedding'] = codebase['summary'].apply(lambda x: generate_api([{'role':'user', 'content': f'将摘要{str(x)}转化为向量'}]))
+    codebase.to_csv(codebase_path, index=True, encoding='utf-8')
+    return f'成功生成摘要向量'
 
 def sum_tokenize(codebase_path):
     codebase = pd.read_csv(codebase_path)
-    codebase['sum_tokenize'] = codebase['summary'].apply(lambda x: jieba.cut_for_search(x))
+    codebase['sum_tokenize'] = codebase['summary'].apply(lambda x: list(jieba.cut_for_search(x)))
+    codebase.to_csv(codebase_path, index=True, encoding='utf-8')
+    return f'成功预分词'
+    
 
 def json_parse(content):
 
@@ -112,25 +125,31 @@ def json_parse(content):
         return None
     return None
 
-def build_code_base(repo_path, lib_path, output_path):
-    
-    if os.path.exists(output_path) and os.path.isdir(output_path):
-        output_path = os.path.join(output_path, f'{os.path.basename(repo_path)}.csv')
-    else:
-        os.makedirs(os.path.dirname(output_path), exist_ok=True)
 
-    if os.path.exists(output_path):
-        result = (f'代码库{os.path.basename(repo_path)}已存在，跳过解析步骤，仅进行数据解析。')
-    else:
-        result = repo_parse(repo_path, lib_path, output_path)
-    gen_sum(codebase_path=output_path)
-    sum_tokenize(codebase_path=output_path)
-    return result
+def force_remove_readonly(func, path, excinfo):
+    """处理只读文件的删除"""
+    os.chmod(path, stat.S_IWRITE)
+    func(path)
 
 def get_repository(repo_url, destination='./repos'):
     if not os.path.exists(destination):
         os.makedirs(destination)
-    print('Start cloning repository...')
-    Repo.clone_from(repo_url, destination)
-    print('Repository cloned to', destination)
-    return destination  
+    destination = os.path.join(destination, repo_url.split('/')[-1].split('.')[0])
+    print(f'开始克隆代码库 {repo_url}')
+    
+    if os.path.exists(destination):
+        rm_repo(destination)
+    repo = Repo.clone_from(repo_url, destination)
+    version = repo.head.commit.hexsha
+    print('代码库克隆到', destination)
+    repo.close()
+    repo.git.clear_cache()
+    return destination, version
+
+def rm_repo(repo_path):
+    for root, dirs, files in os.walk(repo_path):  
+        for dir in dirs:
+            os.chmod(os.path.join(root, dir), stat.S_IRWXU)
+        for file in files:
+            os.chmod(os.path.join(root, file), stat.S_IRWXU)
+    shutil.rmtree(repo_path)
